Refactor by extracting functions and clean up
